Project Started; Heavy Progress
I have started this project and have to close out for the rest of the day, but just like the last one, I feel this one is also easy, if not easier with the added experience.

Additions include the JS link to the file which so far has variables being declared with various text strings and methods; all of which done with documentation along the way.